Reimplement mining of SubTxs and transitions
Initial implementation was lost after merging dashpay/develop into
dashevo/develop. Too much stuff has changed in miner.cpp and thus a
reimplementation is needed.

This implementation is currently a bit limited in regard to which topups
can be included from addPackageTxs. This is because packages currently do
not take the dependency between registers and topups into account. As a
temporary solution addRegisterSubTxs was added which adds all register
SubTxs to the block before we try to add topups in addPackageTxs.
addRegisterSubTxs however ignores fees at the moment (it just adds all)