Also ban masternode when a CProUpRegTx changes the operator key
In that case it's not clear if the owner wants to only switch to a new key
or also switch to another operator. To be secure, we reset all operator
related fields now and ban the MN. A CProUpServTx will be required to
revive it.

Also added two helper methods for resetting operator fields and ban if not
already banned.