Move creating addresses to bind to any interface to network backend.
New virtual method CNetBackend::bind_any_addrs() added to CNetBackend class.
It returns vector of addresses for this backend that can be used
to bind to all local interfaces.

Binding to these addresses in AppInitMain() function in "init.cpp"
changed to use this method instead of using explicit IPv4 and IPv6 addresses.

Signed-off-by: Oleg Girko <ol@infoserver.lv>