fix: adjust EvoDB GetMemoryUsage by factor of 20
During IDB / reindex, memory usage during specific periods would spike as high as 9gb. Especially on nodes with less memory available, this was resulting in bad thrashing and terrible performance. This fixes that, albiet in a less than ideal way. Ideally, we should figure out why we are underestimating the memory usage, however this is a fine fix which will result in purging cache to disk more often. On systems with plenty of memory, this may result in slightly degraded performance during IDB / reindex due to more frequent pushing to disk.