ci: add mirrors to Bordeaux substitution server
Needed to accommodate for differences in network performance due to
geography as each new container instance has to re-download everything.
We don't need to import new public keys as they inherit the key for the
server they're mirroring (i.e. the mirror cannot modify the contents it
is mirroring without _also_ invalidating it), so the risk of malicious
changes are relatively mitigated.